middleware: Moved all classes under setCookie to setData and removed setCookie.py
